move chrome db adapter variable from unit to integration
